refactor: directly use interface type instead of dedicated class
